json file settings backend

+++ core/sqlite-config/src/main/java/org/n52/sos/config/sqlite/hibernate/HibernateSQLiteDialect.java
@@ -86,7 +86,7 @@ public class HibernateSQLiteDialect extends Dialect {
 
     @Override
     public String getIdentityColumnString() {
-        return "integer";
+        return "bigint";
     }
 
     @Override
